feat: implement basic counties map in d3.js
