+++ package.json
@@ -12,6 +12,9 @@
   "dependencies": {
     "@mapbox/polylabel": "^1.0.2",
     "axios": "^1.6.8",
+    "d3": "^7.9.0",
+    "d3-scale": "^4.0.2",
+    "d3-scale-chromatic": "^3.1.0",
     "mapbox-gl": "^3.2.0",
     "vue": "^3.4.21",
     "vue-router": "^4.3.0"
