+++ src/App.vue
@@ -13,6 +13,6 @@ body {
 	height: 100%;
 }
 main {
-	min-height: calc(100vh - 80px - 500px);
+	min-height: calc(100vh - 80px - 60px);
 }
 </style>
