+++ src/pages/HomeRoute.vue
@@ -1,0 +1,8 @@
+<template>
+	PDAP maps
+
+	<div class="flex flex-col">
+		<router-link to="/mapbox">Mapbox map</router-link>
+		<router-link to="/d3">d3.js map</router-link>
+	</div>
+</template>
